+++ cmd/rabtap/message_reader_dir.go
@@ -4,8 +4,13 @@
 package main
 
 import (
+	"fmt"
+	"io/ioutil"
 	"os"
+	"path"
 	"regexp"
+
+	"github.com/streadway/amqp"
 )
 
 const metadataFilePattern = `^rabtap-[0-9]+.json$`
@@ -13,14 +18,14 @@ const metadataFilePattern = `^rabtap-[0-9]+.json$`
 type DirReader func(string) ([]os.FileInfo, error)
 type FileInfoPredicate func(fileinfo os.FileInfo) bool
 
-type filenameWithMetadata struct {
+type FilenameWithMetadata struct {
 	filename string
 	metadata RabtapPersistentMessage
 }
 
 // newRabtapFileInfoPredicate returns a FileInfoPredicate that matches
 // rabtap metadata files
-func newRabtapFileInfoPredicate() FileInfoPredicate {
+func NewRabtapFileInfoPredicate() FileInfoPredicate {
 	filenameRe := regexp.MustCompile(metadataFilePattern)
 	return func(fi os.FileInfo) bool {
 		return fi.Mode().IsRegular() && filenameRe.MatchString(fi.Name())
@@ -37,7 +42,7 @@ func filterMetadataFilenames(fileinfos []os.FileInfo, pred FileInfoPredicate) []
 	return filenames
 }
 
-// findMetadataFilenames returns list of filenames looking like a rabtap
+// findMetadataFilenames returns list of absolute filenames looking like rabtap
 // persisted message/metadata files
 func findMetadataFilenames(dirname string, dirReader DirReader, pred FileInfoPredicate) ([]string, error) {
 	fileinfos, err := dirReader(dirname)
@@ -58,11 +63,12 @@ func readRabtapPersistentMessage(filename string) (RabtapPersistentMessage, erro
 
 // readMetadataOfFiles reads all metadata files from the given list of files.
 // returns an error if any error occurs.
-func readMetadataOfFiles(filenames []string) ([]filenameWithMetadata, error) {
+func readMetadataOfFiles(dirname string, filenames []string) ([]FilenameWithMetadata, error) {
 
-	data := make([]filenameWithMetadata, len(filenames))
+	data := make([]FilenameWithMetadata, len(filenames))
 	for i, filename := range filenames {
-		msg, err := readRabtapPersistentMessage(filename)
+		fullpath := path.Join(dirname, filename)
+		msg, err := readRabtapPersistentMessage(fullpath)
 		if err != nil {
 			return data, err
 		}
@@ -70,18 +76,45 @@ func readMetadataOfFiles(filenames []string) ([]filenameWithMetadata, error) {
 		// JSON or a separate message file). This approach read message bodies
 		// twice, but this should not be a problem (for now...)
 		msg.Body = []byte("")
-		data[i] = filenameWithMetadata{filename: filename, metadata: msg}
+		data[i] = FilenameWithMetadata{filename: fullpath, metadata: msg}
 	}
 
 	return data, nil
 }
 
-// loadMetadataFromDir loads all metadata files from the given directory
+// LoadMetadataFromDir loads all metadata files from the given directory
 // passing the given predicate
-func loadMetadataFilesFromDir(dirname string, dirReader DirReader, pred FileInfoPredicate) ([]filenameWithMetadata, error) {
+func LoadMetadataFilesFromDir(dirname string, dirReader DirReader, pred FileInfoPredicate) ([]FilenameWithMetadata, error) {
 	filenames, err := findMetadataFilenames(dirname, dirReader, pred)
 	if err != nil {
 		return nil, err
 	}
-	return readMetadataOfFiles(filenames)
+	return readMetadataOfFiles(dirname, filenames)
+}
+
+// createMessageFromDirReaderFunc returns a MessageReaderFunc that reads
+// messages from the given list of filenames in the given format.
+func CreateMessageFromDirReaderFunc(format string, files []FilenameWithMetadata) (MessageReaderFunc, error) {
+
+	i := 0
+
+	switch format {
+	case "json-nopp":
+		fallthrough
+	case "json":
+		return func() (amqp.Publishing, bool, error) {
+			fullMessage, err := readRabtapPersistentMessage(files[i].filename)
+			i++
+			return fullMessage.ToAmqpPublishing(), i < len(files), err
+		}, nil
+	case "raw":
+		return func() (amqp.Publishing, bool, error) {
+			body, err := ioutil.ReadFile(files[i].filename)
+			message := files[i].metadata
+			message.Body = body
+			i++
+			return message.ToAmqpPublishing(), i < len(files), err
+		}, nil
+	}
+	return nil, fmt.Errorf("invaild format %s", format)
 }